+++ internal/service/player_service.go
@@ -18,11 +18,11 @@ func(s *PlayerService) GetAll(ctx context.Context) ([]*models.Player, error) {
 	return s.Repo.GetAll(ctx)
 }
 
-func(s *PlayerService) GetPlayersByIDs(ctx context.Context, player_ids []int) ([]*models.Player, error) {
+func(s *PlayerService) GetPlayersByIDs(ctx context.Context, player_ids []int64) ([]*models.Player, error) {
 	return s.Repo.GetByIDs(ctx, player_ids)
 }
 
-func(s *PlayerService) GetPlayerByID(ctx context.Context, id int) (*models.Player, error) {
+func(s *PlayerService) GetPlayerByID(ctx context.Context, id int64) (*models.Player, error) {
 	return s.Repo.GetByID(ctx, id)
 }
 
@@ -35,6 +35,6 @@ func(s *PlayerService) CreatePlayer(ctx context.Context, name string, value floa
 	return s.Repo.Create(ctx, p)
 }
 
-func(s *PlayerService) DeletePlayer(ctx context.Context, id int) (error){
+func(s *PlayerService) DeletePlayer(ctx context.Context, id int64) (error){
 	return s.Repo.Delete(ctx, id)
 }
